TSDK-916 Cleanup
